mv usr appimages to root
appimaged creates desktop launchers using a unique id (based on what?).

The pinned launchers in the dock are "fine" when one uses the default user but are broken when a new user is created, admitedly because the unique id changes.

But, the appimages in roo (/Applications) don't seem to be affected by this.